# openapi.yaml
openapi: 3.0.3
info:
  title: gdl-api
  description: gdl-api express api server + frontend + bots
  version: 0.2.31
  license:
    name: MIT
servers:
  - url: https://gdl.mcalec.dev/api
    description: production server
  - url: http://localhost:3030/api
    description: local server
...
paths:
...
  /auth/login/:
    post:
      tags:
        - Authentication
      summary: User login with username/email and password
      description: Authenticate a user with their credentials (username or email and password)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  description: Username or email of the user
                email:
                  type: string
                  description: Email address of the user (alternative to username)
                password:
                  type: string
                  description: User password
              required:
                - password
      responses:
        "200":
          description: User successfully logged in
        "400":
          description: Invalid request parameters
        "401":
          description: Invalid credentials
        "403":
          description: User already logged in
        "500":
          description: Internal server error
  /auth/logout/:
    post:
      tags:
        - Authentication
      summary: User logout
      description: Log out the authenticated user and destroy their session
      responses:
        "201":
          description: User successfully logged out
        "400":
          description: User is not logged in
        "500":
          description: Internal server error
  /auth/register/:
    post:
      tags:
        - Authentication
      summary: User registration
      description: Register a new user with username, email, and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  description: Unique username for the user
                email:
                  type: string
                  format: email
                  description: Email address of the user
                password:
                  type: string
                  description: Password for the user account
              required:
                - username
                - password
      responses:
        "201":
          description: User successfully registered and logged in
        "400":
          description: Missing required fields or invalid email format
        "409":
          description: Username or email already exists
        "500":
          description: Internal server error
...
  /auth/provider/login/{provider}:
    get:
      tags:
        - Authentication
      summary: Initiate OAuth login
      description: Start OAuth authentication flow for the specified provider
      parameters:
        - in: path
          name: provider
          required: true
          schema:
            type: string
            enum: [github, discord]
          description: OAuth provider name
      responses:
        "302":
          description: Redirect to provider authentication
        "400":
          description: Invalid provider
        "500":
          description: Authentication error
  /auth/provider/callback/{provider}:
    get:
      tags:
        - Authentication
      summary: OAuth callback endpoint
      description: Handle OAuth provider callback after successful authentication
      parameters:
        - in: path
          name: provider
          required: true
          schema:
            type: string
            enum: [github, discord]
          description: OAuth provider name
        - in: query
          name: code
          required: true
          schema:
            type: string
          description: Authorization code from provider
      responses:
        "302":
          description: Redirect to dashboard on success
  /auth/provider/link/{provider}:
    get:
      tags:
        - Authentication
      summary: Link OAuth account to existing user
      description: Link an OAuth provider account to the currently authenticated user
      parameters:
        - in: path
          name: provider
          required: true
          schema:
            type: string
            enum: [github, discord]
          description: OAuth provider name
      responses:
        "302":
          description: Redirect to provider authentication
        "400":
          description: Invalid provider or provider already linked
        "401":
          description: User not authenticated
  /auth/provider/unlink/{provider}:
    get:
      tags:
        - Authentication
      summary: Unlink OAuth account from user
      description: Remove OAuth provider account from the currently authenticated user
      parameters:
        - in: path
          name: provider
          required: true
          schema:
            type: string
            enum: [github, discord]
          description: OAuth provider name
      responses:
        "200":
          description: Provider unlinked successfully
        "400":
          description: Invalid provider or cannot unlink (no other auth method)
        "401":
          description: User not authenticated
...
components:
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: connect.sid
      description: Session cookie for authenticated requests
security:
  - cookieAuth: []